Separate dropdown and item
