+++ src/lib/components/menu/dropdown/dropdown.scss
@@ -1,0 +1,13 @@
+@import '../../../styles/variables';
+
+.menu__dropdown {
+  height: 0;
+  margin: 0;
+  margin-left: $space-32;
+  overflow: hidden;
+  padding: 0;
+
+  &--active {
+    height: auto;
+  }
+}
